+++ src/Symfony/Component/Messenger/Tests/TraceableMessageBusTest.php
@@ -20,20 +20,18 @@ use Symfony\Component\Messenger\TraceableMessageBus;
 
 class TraceableMessageBusTest extends TestCase
 {
-    public function testItTracesResult()
+    public function testItTracesDispatch()
     {
         $message = new DummyMessage('Hello');
 
         $bus = $this->getMockBuilder(MessageBusInterface::class)->getMock();
-        $bus->expects($this->once())->method('dispatch')->with($message)->willReturn($result = array('foo' => 'bar'));
 
         $traceableBus = new TraceableMessageBus($bus);
         $line = __LINE__ + 1;
-        $this->assertSame($result, $traceableBus->dispatch($message));
+        $this->assertNull($traceableBus->dispatch($message));
         $this->assertCount(1, $tracedMessages = $traceableBus->getDispatchedMessages());
         $this->assertArraySubset(array(
             'message' => $message,
-            'result' => $result,
             'stamps' => null,
             'caller' => array(
                 'name' => 'TraceableMessageBusTest.php',
@@ -43,20 +41,18 @@ class TraceableMessageBusTest extends TestCase
         ), $tracedMessages[0], true);
     }
 
-    public function testItTracesResultWithEnvelope()
+    public function testItTracesDispatchWithEnvelope()
     {
         $envelope = Envelope::wrap($message = new DummyMessage('Hello'))->with($stamp = new AnEnvelopeStamp());
 
         $bus = $this->getMockBuilder(MessageBusInterface::class)->getMock();
-        $bus->expects($this->once())->method('dispatch')->with($envelope)->willReturn($result = array('foo' => 'bar'));
 
         $traceableBus = new TraceableMessageBus($bus);
         $line = __LINE__ + 1;
-        $this->assertSame($result, $traceableBus->dispatch($envelope));
+        $this->assertNull($traceableBus->dispatch($envelope));
         $this->assertCount(1, $tracedMessages = $traceableBus->getDispatchedMessages());
         $this->assertArraySubset(array(
             'message' => $message,
-            'result' => $result,
             'stamps' => array($stamp),
             'caller' => array(
                 'name' => 'TraceableMessageBusTest.php',